Conflicts:
	Resources/Core Hacks.txt
	Source/craft/app/assetsourcetypes/GoogleCloudAssetSourceType.php
	Source/craft/app/assetsourcetypes/RackspaceAssetSourceType.php
	Source/craft/app/assetsourcetypes/S3AssetSourceType.php
	Source/craft/app/controllers/AssetsController.php
	Source/craft/app/controllers/DashboardController.php
	Source/craft/app/controllers/TasksController.php
	Source/craft/app/controllers/UpdateController.php
	Source/craft/app/controllers/UsersController.php
	Source/craft/app/elementtypes/AssetElementType.php
	Source/craft/app/elementtypes/CategoryElementType.php
	Source/craft/app/elementtypes/EntryElementType.php
	Source/craft/app/elementtypes/UserElementType.php
	Source/craft/app/etc/behaviors/AppBehavior.php
	Source/craft/app/etc/config/defaults/general.php
	Source/craft/app/etc/db/DbBackup.php
	Source/craft/app/etc/db/DbConnection.php
	Source/craft/app/etc/et/Et.php
	Source/craft/app/etc/i18n/PhpMessageSource.php
	Source/craft/app/etc/requirements/Requirements.php
	Source/craft/app/etc/templating/twigextensions/Cache_Node.php
	Source/craft/app/etc/templating/twigextensions/CraftTwigExtension.php
	Source/craft/app/etc/web/HttpCookie.php
	Source/craft/app/etc/web/WebApp.php
	Source/craft/app/fieldtypes/AssetsFieldType.php
	Source/craft/app/fieldtypes/BaseElementFieldType.php
	Source/craft/app/fieldtypes/BaseFieldType.php
	Source/craft/app/fieldtypes/ColorFieldType.php
	Source/craft/app/fieldtypes/PositionSelectFieldType.php
	Source/craft/app/fieldtypes/RichTextFieldType.php
	Source/craft/app/framework/i18n/data/no.php
	Source/craft/app/helpers/StringHelper.php
	Source/craft/app/helpers/TemplateHelper.php
	Source/craft/app/lib/GC.php
	Source/craft/app/lib/S3.php
	Source/craft/app/models/AssetFileModel.php
	Source/craft/app/models/BaseElementModel.php
	Source/craft/app/models/ElementCriteriaModel.php
	Source/craft/app/models/UserModel.php
	Source/craft/app/records/FieldRecord.php
	Source/craft/app/records/TagGroupRecord.php
	Source/craft/app/records/UserRecord.php
	Source/craft/app/resources/css/source/craft.scss
	Source/craft/app/resources/js/asseteditsourcesettings.js
	Source/craft/app/resources/js/classes/AssetIndex.js
	Source/craft/app/resources/js/classes/AuthManager.js
	Source/craft/app/resources/js/classes/BaseElementIndex.js
	Source/craft/app/resources/js/classes/EntryIndex.js
	Source/craft/app/resources/js/classes/LivePreview.js
	Source/craft/app/resources/js/classes/StructureTableSorter.js
	Source/craft/app/resources/js/cp.js
	Source/craft/app/resources/js/craft.js
	Source/craft/app/resources/lib/redactor/redactor.js
	Source/craft/app/resources/lib/redactor/redactor.min.js
	Source/craft/app/services/AssetSourcesService.php
	Source/craft/app/services/AssetTransformsService.php
	Source/craft/app/services/AssetsService.php
	Source/craft/app/services/CacheService.php
	Source/craft/app/services/CategoriesService.php
	Source/craft/app/services/ConfigService.php
	Source/craft/app/services/ElementsService.php
	Source/craft/app/services/EmailService.php
	Source/craft/app/services/EntriesService.php
	Source/craft/app/services/FeedsService.php
	Source/craft/app/services/HttpRequestService.php
	Source/craft/app/services/HttpSessionService.php
	Source/craft/app/services/ImagesService.php
	Source/craft/app/services/LocalizationService.php
	Source/craft/app/services/PluginsService.php
	Source/craft/app/services/SectionsService.php
	Source/craft/app/services/TasksService.php
	Source/craft/app/services/TemplateCacheService.php
	Source/craft/app/services/TemplatesService.php
	Source/craft/app/services/TokensService.php
	Source/craft/app/services/UserSessionService.php
	Source/craft/app/services/UsersService.php
	Source/craft/app/tasks/DeleteStaleTemplateCachesTask.php
	Source/craft/app/templates/_components/assetsourcetypes/GoogleCloud/settings.html
	Source/craft/app/templates/_components/assetsourcetypes/Rackspace/settings.html
	Source/craft/app/templates/_components/assetsourcetypes/S3/settings.html
	Source/craft/app/templates/updates/index.html
	Source/craft/app/tools/DbBackupTool.php
	Source/craft/app/translations/de.php
	Source/craft/app/translations/en_gb.php
	Source/craft/app/translations/en_us.php
	Source/craft/app/translations/es.php
	Source/craft/app/translations/fr.php
	Source/craft/app/translations/fr_ca.php
	Source/craft/app/translations/he.php
	Source/craft/app/translations/it.php
	Source/craft/app/translations/ja.php
	Source/craft/app/translations/ko.php
	Source/craft/app/translations/nl.php
	Source/craft/app/translations/no.php
	Source/craft/app/translations/sv.php
	Source/craft/app/translations/zh.php
	Source/craft/app/variables/CpVariable.php
	Source/craft/app/variables/CraftVariable.php
	Source/craft/app/variables/RoutesVariable.php
	Source/craft/app/vendor/lsolesen/pel/src/PelEntryAscii.php
	Utils/Builder/Builder.php

+++ src/services/Assets.php
@@ -461,15 +461,17 @@ class Assets extends Component
 				));
 				$this->trigger(static::EVENT_BEFORE_DELETE_ASSET, $event);
 
-				if ($deleteFile)
+				if ($event->performAction)
 				{
-					$volume->deleteFile($file->getUri());
+					if ($deleteFile)
+					{
+						$volume->deleteFile($file->getUri());
+					}
+
+					Craft::$app->getElements()->deleteElementById($fileId);
+
+					$this->trigger(static::EVENT_AFTER_DELETE_ASSET, $event);
 				}
-
-				Craft::$app->getElements()->deleteElementById($fileId);
-
-				$this->trigger(static::EVENT_AFTER_DELETE_ASSET, $event);
-
 			}
 		}
 	}
